Split task code into multiple tasks files
