+++ packages/aragon-cli/src/commands/apm_cmds/publish.js
@@ -3,36 +3,20 @@ const tmp = require('tmp-promise')
 const path = require('path')
 const { readJson, writeJson, pathExistsSync } = require('fs-extra')
 const APM = require('@aragon/apm')
-const semver = require('semver')
 const TaskList = require('listr')
 const taskInput = require('listr-input')
 const inquirer = require('inquirer')
 const chalk = require('chalk')
-const { findProjectRoot, runScriptTask, ZERO_ADDRESS } = require('../../util')
-const { compileContracts } = require('../../helpers/truffle-runner')
-const web3Utils = require('web3').utils
+const { findProjectRoot, ZERO_ADDRESS } = require('../../util')
 const deploy = require('../deploy')
-const startIPFS = require('../ipfs_cmds/start')
 const propagateIPFS = require('../ipfs_cmds/propagate')
 const execTask = require('../dao_cmds/utils/execHandler').task
 const listrOpts = require('@aragon/cli-utils/src/helpers/listr-options')
 
-const {
-  prepareFilesForPublishing,
-  MANIFEST_FILE,
-  ARTIFACT_FILE,
-} = require('./util/preprare-files')
-
-const {
-  getMajor,
-  sanityCheck,
-  generateApplicationArtifact,
-  generateFlattenedCode,
-  copyCurrentApplicationArtifacts,
-  SOLIDITY_FILE,
-  POSITIVE_ANSWERS,
-  ANSWERS,
-} = require('./util/generate-artifact')
+// Tasks splitted into other files
+const runSetupTask = require('./publish/runSetupTask')
+const runPrepareForPublishTask = require('./publish/runPrepareForPublishTask')
+const runPublishTask = require('./publish/runPublishTask')
 
 exports.command = 'publish <bump> [contract]'
 
@@ -140,433 +124,10 @@ exports.builder = function(yargs) {
     })
 }
 
-exports.runSetupTask = ({
-  reporter,
-
-  // Globals
-  gasPrice,
-  cwd,
-  web3,
-  network,
-  module,
-  apm: apmOptions,
-  silent,
-  debug,
-
-  // Arguments
-
-  /// Scritps
-  prepublish,
-  prepublishScript,
-  build,
-  buildScript,
-
-  /// Version
-  bump,
-
-  /// Contract
-  contract,
-  init,
-  reuse,
-
-  /// Conditionals
-  onlyContent,
-  onlyArtifacts,
-  ipfsCheck,
-  http,
-}) => {
-  if (onlyContent) {
-    contract = ZERO_ADDRESS
-  }
-  apmOptions.ensRegistryAddress = apmOptions['ens-registry']
-  const apm = APM(web3, apmOptions)
-
-  return new TaskList(
-    [
-      {
-        title: 'Running prepublish script',
-        enabled: () => prepublish,
-        task: async (ctx, task) => runScriptTask(task, prepublishScript),
-      },
-      {
-        title: 'Check IPFS',
-        enabled: () => !http && ipfsCheck,
-        task: () => startIPFS.task({ apmOptions }),
-      },
-      {
-        title: `Applying version bump (${bump})`,
-        task: async ctx => {
-          let isValid = true
-          try {
-            const ipfsTimeout = 1000 * 60 * 5 // 5min
-            reporter.info('Fetching latest version from aragonPM...')
-            ctx.initialRepo = await apm.getLatestVersion(
-              module.appName,
-              ipfsTimeout
-            )
-
-            ctx.initialVersion = ctx.initialRepo.version
-
-            ctx.version = semver.valid(bump)
-              ? semver.valid(bump)
-              : semver.inc(ctx.initialVersion, bump)
-
-            isValid = await apm.isValidBump(
-              module.appName,
-              ctx.initialVersion,
-              ctx.version
-            )
-            if (!isValid) {
-              throw new Error(
-                "Version bump is not valid, you have to respect APM's versioning policy. Check the version upgrade rules in the documentation: https://hack.aragon.org/docs/apm-ref.html#version-upgrade-rules"
-              )
-            }
-
-            ctx.shouldDeployContract =
-              getMajor(ctx.initialVersion) !== getMajor(ctx.version)
-          } catch (e) {
-            if (e.message.indexOf('Invalid content URI') === 0) {
-              return
-            }
-            // Repo doesn't exist yet, deploy the first version
-            ctx.version = semver.valid(bump)
-              ? semver.valid(bump)
-              : semver.inc('0.0.0', bump) // All valid initial versions are a version bump from 0.0.0
-            if (apm.validInitialVersions.indexOf(ctx.version) === -1) {
-              throw new Error(
-                `Invalid initial version  (${ctx.version}). It can only be 0.0.1, 0.1.0 or 1.0.0.`
-              )
-            }
-            ctx.shouldDeployContract = true // assume first version should deploy a contract
-          }
-        },
-      },
-      {
-        title: 'Building frontend',
-        enabled: () => build && !http,
-        task: async (ctx, task) => runScriptTask(task, buildScript),
-      },
-      {
-        title: 'Compile contracts',
-        enabled: () => !onlyContent && web3Utils.isAddress(contract),
-        task: async () => compileContracts(),
-      },
-      {
-        title: 'Deploy contract',
-        enabled: ctx =>
-          !onlyContent &&
-          ((contract && !web3Utils.isAddress(contract)) ||
-            (!contract && ctx.shouldDeployContract && !reuse)),
-        task: async ctx => {
-          const deployTaskParams = {
-            module,
-            contract,
-            init,
-            gasPrice,
-            network,
-            cwd,
-            web3,
-            apmOptions,
-          }
-
-          return deploy.task(deployTaskParams)
-        },
-      },
-      {
-        title: 'Determine contract address for version',
-        enabled: () => !onlyArtifacts,
-        task: async (ctx, task) => {
-          if (web3Utils.isAddress(contract)) {
-            ctx.contract = contract
-          }
-
-          // Check if we can fall back to a previous contract address
-          if (
-            !ctx.contract &&
-            apm.validInitialVersions.indexOf(ctx.version) === -1
-          ) {
-            task.output = 'No contract address provided, using previous one'
-
-            try {
-              const { contractAddress } = await apm.getLatestVersion(
-                module.appName
-              )
-              ctx.contract = contractAddress
-              return `Using ${ctx.contract}`
-            } catch (err) {
-              throw new Error('Could not determine previous contract')
-            }
-          }
-
-          // Contract address required for initial version
-          if (!ctx.contract) {
-            throw new Error('No contract address supplied for initial version')
-          }
-
-          return `Using ${contract}`
-        },
-      },
-    ],
-    listrOpts(silent, debug)
-  )
-}
-
-exports.runPrepareForPublishTask = ({
-  reporter,
-
-  // Globals
-  cwd,
-  web3,
-  network,
-  module,
-  apm: apmOptions,
-  silent,
-  debug,
-
-  // Arguments
-
-  /// Files
-  publishDir,
-  files,
-  ignore,
-
-  /// Http
-  httpServedFrom,
-
-  /// Storage
-  provider,
-
-  /// Conditionals
-  onlyArtifacts,
-  onlyContent,
-  http,
-
-  // Context
-  initialRepo,
-  initialVersion,
-  version,
-  contractAddress,
-  deployArtifacts,
-}) => {
-  apmOptions.ensRegistryAddress = apmOptions['ens-registry']
-  const apm = APM(web3, apmOptions)
-
-  return new TaskList(
-    [
-      {
-        title: 'Prepare files for publishing',
-        enabled: () => !http,
-        task: async (ctx, task) => {
-          // Create temporary directory
-          if (!publishDir) {
-            const { path: tmpDir } = await tmp.dir()
-            publishDir = tmpDir
-          }
-
-          await prepareFilesForPublishing(publishDir, files, ignore)
-          ctx.pathToPublish = publishDir
-
-          return `Files copied to temporary directory: ${ctx.pathToPublish}`
-        },
-      },
-      {
-        title:
-          'Check for --http-served-from argument and copy manifest.json to destination',
-        enabled: () => http,
-        task: async (ctx, task) => {
-          if (!httpServedFrom) {
-            throw new Error('You need to provide --http-served-from argument')
-          }
-
-          const projectRoot = findProjectRoot()
-          const manifestOrigin = path.resolve(projectRoot, MANIFEST_FILE)
-          const manifestDst = path.resolve(httpServedFrom, MANIFEST_FILE)
-
-          if (!pathExistsSync(manifestDst) && pathExistsSync(manifestOrigin)) {
-            let manifest = await readJson(manifestOrigin)
-            manifest.start_url = path.basename(manifest.start_url)
-            manifest.script = path.basename(manifest.script)
-            await writeJson(manifestDst, manifest)
-          }
-
-          ctx.pathToPublish = httpServedFrom
-        },
-      },
-      {
-        title: 'Generate application artifact',
-        skip: () => onlyContent && !module.path,
-        task: async (ctx, task) => {
-          const dir = onlyArtifacts ? cwd : ctx.pathToPublish
-
-          const contractPath = module.path
-          const roles = module.roles
-
-          // TODO: (Gabi) Use inquier to handle confirmation
-          async function invokeArtifactGeneration(answer) {
-            if (POSITIVE_ANSWERS.indexOf(answer) > -1) {
-              await generateApplicationArtifact(
-                cwd,
-                apm,
-                dir,
-                module,
-                deployArtifacts,
-                web3,
-                reporter
-              )
-              await generateFlattenedCode(dir, contractPath)
-              return `Saved artifact in ${dir}/${ARTIFACT_FILE}`
-            }
-            throw new Error('Aborting publication...')
-          }
-
-          // If an artifact file exist we check it to reuse
-          if (pathExistsSync(`${dir}/${ARTIFACT_FILE}`)) {
-            const existingArtifactPath = path.resolve(dir, ARTIFACT_FILE)
-            const existingArtifact = await readJson(existingArtifactPath)
-            const rebuild = await sanityCheck(
-              cwd,
-              roles,
-              contractPath,
-              existingArtifact
-            )
-            if (rebuild) {
-              return taskInput(
-                `Couldn't reuse artifact due to mismatches, regenerate now? [y]es/[a]bort`,
-                {
-                  validate: value => {
-                    return ANSWERS.indexOf(value) > -1
-                  },
-                  done: async answer => invokeArtifactGeneration(answer),
-                }
-              )
-            } else {
-              return task.skip('Using existing artifact')
-            }
-          }
-
-          // If only content we fetch artifacts from previous version
-          if (
-            onlyContent &
-            (apm.validInitialVersions.indexOf(version) === -1)
-          ) {
-            try {
-              task.output = 'Fetching artifacts from previous version'
-              await copyCurrentApplicationArtifacts(
-                cwd,
-                dir,
-                apm,
-                initialRepo,
-                version,
-                roles,
-                contractPath
-              )
-              if (!pathExistsSync(`${dir}/${SOLIDITY_FILE}`)) {
-                await generateFlattenedCode(dir, contractPath)
-              }
-              return task.skip(`Using artifacts from v${initialVersion}`)
-            } catch (e) {
-              if (e.message === 'Artifact mismatch') {
-                return taskInput(
-                  "Couldn't reuse existing artifact due to mismatches, regenerate now? [y]es/[a]bort",
-                  {
-                    validate: value => {
-                      return ANSWERS.indexOf(value) > -1
-                    },
-                    done: async answer => invokeArtifactGeneration(answer),
-                  }
-                )
-              } else {
-                return taskInput(
-                  "Couldn't fetch current artifact version to copy it. Please make sure your IPFS or HTTP server are running. Otherwise, generate now? [y]es/[a]bort",
-                  {
-                    validate: value => {
-                      return ANSWERS.indexOf(value) > -1
-                    },
-                    done: async answer => invokeArtifactGeneration(answer),
-                  }
-                )
-              }
-            }
-          }
-
-          await invokeArtifactGeneration('yes')
-
-          return `Saved artifact in ${dir}/artifact.json`
-        },
-      },
-      {
-        title: `Publish intent`,
-        task: async (ctx, task) => {
-          ctx.contractInstance = null // clean up deploy sub-command artifacts
-          const accounts = await web3.eth.getAccounts()
-          const from = accounts[0]
-          ctx.intent = await apm.publishVersionIntent(
-            from,
-            module.appName,
-            version,
-            http ? 'http' : provider,
-            http || ctx.pathToPublish,
-            contractAddress
-          )
-        },
-      },
-    ],
-    listrOpts(silent, debug)
-  )
-}
-
-exports.runPublishTask = ({
-  reporter,
-
-  // Globals
-  gasPrice,
-  web3,
-  wsProvider,
-  module,
-  apm: apmOptions,
-  silent,
-  debug,
-
-  // Arguments
-  /// Conditionals
-  onlyArtifacts,
-  onlyContent,
-
-  /// Context
-  dao,
-  proxyAddress,
-  methodName,
-  params,
-}) => {
-  apmOptions.ensRegistryAddress = apmOptions['ens-registry']
-  return new TaskList(
-    [
-      {
-        title: `Publish ${module.appName}`,
-        enabled: () => !onlyArtifacts,
-        task: async (ctx, task) => {
-          try {
-            return execTask({
-              dao,
-              app: proxyAddress,
-              method: methodName,
-              params,
-              ipfsCheck: false,
-              reporter,
-              gasPrice,
-              apm: apmOptions,
-              web3,
-              wsProvider,
-            })
-          } catch (e) {
-            throw e
-          }
-        },
-      },
-    ],
-    listrOpts(silent, debug)
-  )
-}
+// Re-export for compatibility with depedant modules
+exports.runSetupTask = runSetupTask
+exports.runPrepareForPublishTask = runPrepareForPublishTask
+exports.runPublishTask = runPublishTask
 
 exports.handler = async function({
   reporter,
@@ -698,9 +259,7 @@ exports.handler = async function({
       '\n',
       'Explore the ipfs content locally:',
       '\n',
-      chalk.bold(
-        `http://localhost:8080/ipfs/QmSDgpiHco5yXdyVTfhKxr3aiJ82ynz8V14QcGKicM3rVh/#/explore/${contentLocation}`
-      ),
+      chalk.bold(`http://localhost:5001/webui/#/explore/${contentLocation}`),
       '\n'
     )
   }
