+++ packages/aragon-cli/src/params.js
@@ -1,0 +1,9 @@
+// Unique common file to specify aragon cli specific constants
+
+const MANIFEST_FILE = 'manifest.json'
+const ARTIFACT_FILE = 'artifact.json'
+
+module.exports = {
+  MANIFEST_FILE,
+  ARTIFACT_FILE,
+}
